Added a script (genupd.R) that builds a country/region list from Spectrum's GBModData file, creates UPD files for each, then uses the DemProjR package to compare population projections to WPP 2024.
Changed to binary file mode when writing UPD files so that writeChar would not throw warnings when writing the UTF-8 BOM. The UTF-8 BOM is written manually so that Excel will interpret UPD files as CSV files on read. The BOM can cause headaches in other processes, so it would be good to find a better solution for this if possible.